requirements.txt
# Core Dependencies
Flask==2.3.2
numpy==1.24.3
scikit-learn==1.2.2
nltk==3.8.1
spacy==3.5.2
chardet==5.1.0
textdistance==4.5.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
langdetect==1.0.9
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
transformers==4.30.2
tensorflow==2.12.0
tf-keras==2.18.0

# File Handling
PyPDF2==3.0.1
python-docx==0.8.11

# Web and Deployment
gunicorn==20.1.0
Werkzeug==2.3.4

# Development and Debugging
pytest==7.3.1
coverage==7.2.5
logging==0.5.1.2

# Optional: For advanced logging
python-json-logger==2.0.7
